Use IncrementalRLSetting in pettingzoo script so that task_labels_at_test_time is set to false (by default with this setting)
